chore: update view_props types
